* master: (220 commits)
  FLUID-6213: Remove corruption when instantiating dynamic component with directly specified subcomponents
  FLUID-6202: Anti-Zalgoisation in fluid.fetchResources to avoid construction races in event binding code
  FLUID-6202: Non-failing test for wrong diagnosis of cause of issue
  FLUID-5534: Updating so that the tests will pass after an npm install.
  FLUID-5534: Updating namespace for separated panel tests.
  FLUID-5534: Adding comment about requiring two iframes in test.
  FLUID-5534: Improving namespace for responsive tests.
  FLUID-5534: Simplifying model by using panelIndex directly in transform
  FLUID-6196: Passing reference to paths around instead
  FLUID-5534: Fixed bug in triggerDOMEvent function
  FLUID-5534: Using sequence injection to setup tests.
  FLUID-5534: Fixing event dispatch for IE11
  FLUID-5534: Adding iframe setup to tests. Now run in all-tests.
  FLUID-6195: Catch block in deenlistModelComponent to prevent permanent framework corruption through exploding relay rule
  FLUID-5534: Added a test for the window resize event.
  FLUID-6194: Removed now unnecessary commented guard - the inverse of a transform may flip from existing to nonexisting as it is edited (implausible, but in theory possible with the current architecture)
  FLUID-6194: Avoid relaying DELETEs backwards for transformations which do not have inverses
  FLUID-6192: Updating copyright years following review
  FLUID-5534: removing window even onDestroy
  FLUID-5534: Uncommenting the beforeReset event defintion
  ...

+++ src/framework/core/js/Fluid.js
@@ -189,7 +189,7 @@ var fluid = fluid || fluid_3_0_0;
     };
 
     fluid.renderLoggingArg = function (arg) {
-        return fluid.isPrimitive(arg) || !fluid.isPlainObject(arg) ? arg : JSON.stringify(arg);
+        return arg === undefined ? "undefined" : fluid.isPrimitive(arg) || !fluid.isPlainObject(arg) ? arg : JSON.stringify(arg);
     };
 
     // The framework's built-in "fail" failure handler - this throws an exception of type <code>fluid.FluidError</code>
@@ -683,22 +683,23 @@ var fluid = fluid || fluid_3_0_0;
         return fluid.filterKeys(toCensor, keys, true);
     };
 
-    // TODO: This is not as clever an idea as we think it is - this typically inner-loop function will optimise badly due to closure
-    fluid.makeFlatten = function (index) {
-        return function (obj) {
-            var togo = [];
-            fluid.each(obj, function (/* value, key */) {
-                togo.push(arguments[index]);
-            });
-            return togo;
-        };
+    /** Return the keys in the supplied object as an array. Note that this will return keys found in the prototype chain as well as "own properties", unlike Object.keys() **/
+    fluid.keys = function (obj) {
+        var togo = [];
+        for (var key in obj) {
+            togo.push(key);
+        }
+        return togo;
     };
 
-    /** Return the keys in the supplied object as an array. Note that this will return keys found in the prototype chain as well as "own properties", unlike Object.keys() **/
-    fluid.keys = fluid.makeFlatten(1);
-
     /** Return the values in the supplied object as an array **/
-    fluid.values = fluid.makeFlatten(0);
+    fluid.values = function (obj) {
+        var togo = [];
+        for (var key in obj) {
+            togo.push(obj[key]);
+        }
+        return togo;
+    };
 
     /**
      * Searches through the supplied object, and returns <code>true</code> if the supplied value
@@ -831,7 +832,7 @@ var fluid = fluid || fluid_3_0_0;
      * Derived from Sindre Sorhus's round-to node module ( https://github.com/sindresorhus/round-to ).
      * License: MIT
      *
-     * Rounds the supplied number to at most the number of decimal places indicated by the scale, omiting any trailing 0s.
+     * Rounds the supplied number to at most the number of decimal places indicated by the scale, omitting any trailing 0s.
      * There are three possible rounding methods described below: "round", "ceil", "floor"
      * Round: Numbers are rounded away from 0 (i.e 0.5 -> 1, -0.5 -> -1).
      * Ceil: Numbers are rounded up
@@ -2375,7 +2376,9 @@ var fluid = fluid || fluid_3_0_0;
         mergeOptions.updateBlocks = updateBlocks;
         mergeOptions.destroyValue = function (segs) { // This method is a temporary hack to assist FLUID-5091
             for (var i = 0; i < mergeBlocks.length; ++i) {
-                fluid.destroyValue(mergeBlocks[i].target, segs);
+                if (!mergeBlocks[i].immutableTarget) {
+                    fluid.destroyValue(mergeBlocks[i].target, segs);
+                }
             }
             fluid.destroyValue(baseMergeOptions.target, segs);
         };
